Fixed the movement of files

+++ clear.py
@@ -20,12 +20,14 @@ def create_shortcut(filename, dest):
 def move_file(filename, start_dest, end_dest):
   counter = 1
   file_exists = os.path.isfile(f'{end_dest}{filename}')
-  while file_exists:
-    new_filename = f'({counter}) - {filename}'
-    counter += 1
-    file_exists = os.path.isfile(f'{end_dest}{new_filename}')
-  # TODO make it work also if there is no need for new_file (when file_exists == false)
-  os.rename(f'{start_dest}{filename}', f'{end_dest}{new_filename}')
+  if file_exists:
+    while file_exists:
+      new_filename = f'({counter}) - {filename}'
+      counter += 1
+      file_exists = os.path.isfile(f'{end_dest}{new_filename}')
+    os.rename(f'{start_dest}{filename}', f'{end_dest}{new_filename}')
+  else:
+    os.rename(f'{start_dest}{filename}', f'{end_dest}{filename}')
 
 
 for filename in os.listdir(directory):
